Security

+++ src/main/java/ro/societateahermes/backendservice/controller/controllerImplementation/NewsFeedController.java
@@ -2,7 +2,10 @@ package ro.societateahermes.backendservice.controller.controllerImplementation;
 
 import org.springframework.web.bind.annotation.*;
 import ro.societateahermes.backendservice.entities.dto.NewsFeedDTO;
+import ro.societateahermes.backendservice.exceptions.UnathorizeException;
 import ro.societateahermes.backendservice.service.serviceImplementation.NewsFeedServiceImplementation;
+import ro.societateahermes.backendservice.utils.PermissionChecker;
+import ro.societateahermes.backendservice.utils.RolesActiveUser;
 
 import java.util.List;
 
@@ -13,21 +16,30 @@ public class NewsFeedController {
 
     private final NewsFeedServiceImplementation newsFeed;
 
-    public NewsFeedController(NewsFeedServiceImplementation newsFeed){
+    public NewsFeedController(NewsFeedServiceImplementation newsFeed) {
         this.newsFeed = newsFeed;
     }
 
     @GetMapping
-    public List<NewsFeedDTO> getNews(){
+    public List<NewsFeedDTO> getNews() {
         return newsFeed.getAllPost();
     }
 
-    @PostMapping
-    public void saveEventToNewsFeed(@RequestBody NewsFeedDTO post){
+    @PostMapping("/{eventId}")
+    public void saveEventToNewsFeed(@PathVariable("eventId") long eventId, @RequestBody NewsFeedDTO post) throws UnathorizeException {
+        List<String> roles = RolesActiveUser.getRoles();
+        if (!PermissionChecker.check(eventId, roles)) {
+            throw new UnathorizeException("User is not authorized");
+        }
         newsFeed.createPost(post);
     }
-    @DeleteMapping("/{newsId}")
-    public void deleteEventFromNewsFeed(@PathVariable("newsId")Long id){
+
+    @DeleteMapping("/{eventId}/{newsId}")
+    public void deleteEventFromNewsFeed(@PathVariable("eventId") long eventId, @PathVariable("newsId") Long id) throws UnathorizeException {
+        List<String> roles = RolesActiveUser.getRoles();
+        if (!PermissionChecker.check(eventId, roles)) {
+            throw new UnathorizeException("User is not authorized");
+        }
         newsFeed.deletePost(id);
     }
 }
