+++ src/main/java/ro/societateahermes/backendservice/controller/controllerImplementation/OfferController.java
@@ -5,7 +5,10 @@ import org.springframework.web.bind.annotation.*;
 import ro.societateahermes.backendservice.entities.Offer;
 import ro.societateahermes.backendservice.entities.dto.FormDto;
 import ro.societateahermes.backendservice.entities.dto.OfferDTO;
+import ro.societateahermes.backendservice.exceptions.UnathorizeException;
 import ro.societateahermes.backendservice.service.OfferServiceInterface;
+import ro.societateahermes.backendservice.utils.PermissionChecker;
+import ro.societateahermes.backendservice.utils.RolesActiveUser;
 
 import java.util.List;
 
@@ -19,24 +22,39 @@ public class OfferController {
     public OfferController(OfferServiceInterface offerService) {
         this.offerService = offerService;
     }
+
     @GetMapping
     public List<OfferDTO> getAll() {
         return offerService.getAll();
     }
 
 
-    @PostMapping
-    public void save(@RequestBody OfferDTO offerDTO) {
+    @PostMapping("/{eventId}")
+    public void save(@PathVariable("eventId") long eventId, @RequestBody OfferDTO offerDTO) throws UnathorizeException {
+        List<String> roles = RolesActiveUser.getRoles();
+        if (!PermissionChecker.check(eventId, roles)) {
+            throw new UnathorizeException("User is not authorized");
+        }
         offerService.save(offerDTO);
     }
 
-    @DeleteMapping("/{offerId}")
-    public void delete(@PathVariable("offerId") Long offerId) {
+    @DeleteMapping("/{eventId}/{offerId}")
+    public void delete(@PathVariable("eventId") long eventId, @PathVariable("offerId") Long offerId) throws UnathorizeException {
+        List<String> roles = RolesActiveUser.getRoles();
+        if (!PermissionChecker.check(eventId, roles)) {
+            throw new UnathorizeException("User is not authorized");
+        }
         offerService.delete(offerId);
     }
 
-    @PutMapping
-    public void update(@RequestBody OfferDTO offerDTO) { offerService.update(offerDTO); }
+    @PutMapping("/{eventId}")
+    public void update(@PathVariable("eventId") long eventId, @RequestBody OfferDTO offerDTO) throws UnathorizeException {
+        List<String> roles = RolesActiveUser.getRoles();
+        if (!PermissionChecker.check(eventId, roles)) {
+            throw new UnathorizeException("User is not authorized");
+        }
+        offerService.update(offerDTO);
+    }
 
     @GetMapping("/{offerId}")
     public OfferDTO getOne(@PathVariable("offerId") long offerId) {
